+++ config.php
@@ -3,7 +3,7 @@
 
 define('SERVER_ROOT', $_SERVER['DOCUMENT_ROOT']);
 
-$config['version']              = '1.8.9';
+$config['version']              = '1.8.11';
 $config['page_prefix']          = ''; // uri cesta mezi domenou a adresarem bistra
 $config['page_free']            = ['', 'favicon.ico']; // to be removed
 $config['folder_backup']        = SERVER_ROOT.'/files/backups/';
@@ -17,3 +17,4 @@ $config['folder_cache']         = SERVER_ROOT.'/cache/'; // Latte cache
 $config['mime-image']           = ['image/jpeg', 'image/pjpeg', 'image/png'];
 $config['platformConfig']       = SERVER_ROOT.'/.env.php';
 $config['dbpass']               = SERVER_ROOT.'/inc/important.php'; // database password - second line of the file
+$config['session_lenght']       = [30,3600]; //user customizable session lenght