+++ lib/case.php
@@ -32,3 +32,28 @@ function casesAssignedTo($userid): array
 //vypis pripadu musi oznacovat NEW - if (@$filter['new'] == null || ($filter['new'] == on && searchRecord(3,$rec['id']))) {
 //pridat sloupec pro created
 //prejmenovat datum za edited
+
+function listCases($closed=null)
+{
+    global $user,$database;
+    if (!$closed) {
+        $sqlFilter = DB_PREFIX.'case.status = 0 AND ';
+    }
+    $sqlFilter .= DB_PREFIX.'case.deleted <= '.$user['aclRoot'].' AND '.DB_PREFIX.'case.secret <= '.$user['aclSecret'];
+    echo   $caseListSql = 'SELECT
+        '.DB_PREFIX.'case.*
+        FROM '.DB_PREFIX.'case
+        LEFT JOIN '.DB_PREFIX.'c2s as caseSolver on '.DB_PREFIX.'case.id = caseSolver.idcase
+        LEFT JOIN '.DB_PREFIX.'user as caseSolverUser on caseSolver.iduser = caseSolverUser.userId
+        LEFT JOIN '.DB_PREFIX.'person as caseSolverPerson on caseSolverUser.personId = caseSolverPerson.id AND caseSolverPerson.deleted = 0 AND caseSolverPerson.secret <= '.$user['aclSecret'].'
+        WHERE '.$sqlFilter;
+    $caseList = mysqli_query($database, $caseListSql);
+    if (mysqli_num_rows($caseList) > 0) {
+        while ($case = mysqli_fetch_assoc($caseList)) {
+            $cases[] = $case;
+        }
+        return $cases;
+    } else {
+        return false;
+    }
+}