+++ sql/update-1.8.11.php
@@ -1,0 +1,15 @@
+<?php
+
+/*
+* DROP COLUMN
+*/
+$columnDrop['report'][] = 'datum';
+$columnDrop['report'][] = 'task';
+$columnDrop['report'][] = 'summary';
+$columnDrop['report'][] = 'impacts';
+$columnDrop['report'][] = 'details';
+$columnDrop['report'][] = 'deleted';
+$columnDrop['report'][] = 'adatum';
+$columnDrop['report'][] = 'energy';
+$columnDrop['report'][] = 'inputs';
+$columnDrop['user'][] = 'plan';